Extract and use package's version for '--version'
It appears pbr is capable of managing versions using Git tags. As a
result it probably doesn't make sense to use a VERSION constant as
this will likely get out of sync rather quickly. Instead, retrieve
the version of the package itself using pip's 'pkg_resources'. This
has the downside of not working when the package is not installed but
this is not the usual use case.

Signed-off-by: Stephen Finucane <stephenfinucane@hotmail.com>