Graceful degradation of patchwork commands
As patchwork steadily adds new functionality, it is inevitable that some
deployments will not be upgraded as fast as others. Rather than preventing
users working with older versions of patchwork from using these servers,
it should be possible to ensure that the features that are supported are
enabled while those that are not disabled. To this end, add support for
checks that ensure the version of patchwork is appropriate for the given
commands.

Signed-off-by: Stephen Finucane <stephenfinucane@hotmail.com>