Reboot!
Time to reboot development of git-pw using the new fangled REST API.
There are a couple of changes involved here, in what is pretty much a
rewrite:

- Use the REST API instead of the XML-RPC API (*)
- Use the 'click' library instead of 'argparse'
- Use subcommands correctly, instead of copying 'git-review's abuse of
  opts (e.g. use 'apply' instead of '-d'). People can create aliases if
  necessary.

The following commands are basically implemented, though there is still
work to be done with logging etc.

- apply
- download

The following commands are stubbed out:

- show
- update
- list

(*) We're not actually using any of the REST API yet, but we will.

Signed-off-by: Stephen Finucane <stephen@that.guru>