Save to file by default
Usability testing by a colleague identified this as a confusing aspect
of the UX. For most people, download roughly equates to "save to a
file", not "output to stdout". Resolve this by downloading to a file
automatically unless specifically requests. This also requires bumping
the logging level to INFO to allow us to output helpful information.

Signed-off-by: Stephen Finucane <stephen@that.guru>