Prepare for required server version
We want to introduce server version-dependent functionality in git-pw.
While using the versionless URL might be slightly shorter, it does make
things a lot tougher to figure out. Start warning users if the server
URL provided doesn't come with a server version. This can be an error in
future versions.

Signed-off-by: Stephen Finucane <stephen@that.guru>