Don't require escaping arguments
It's helpful being able to pass through additional arguments to
'git-am', where it's used. We've supported this since 1.0 but we needed
to escape these arguments like so:

  git-am apply 123 -- --signoff

This is, as seen in #19 and #20, not very intuitive. Remove the need to
do this.

  git-am apply 123 --signoff

Note that some arguments - notably '--help' - may need to be escaped as
before to prevent our own parser swallowing them.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Closes: #20