tests: Validate unescaped arguments
In commit 6f045bb8, we modified the various 'git-am' wrappers so that
additional arguments to be passed to 'git-am' did not need to be
escaped, such that:

  git-am apply 123 -- --signoff

would become:

  git-am apply 123 --signoff

Modify the unit tests for these commands to validate this behavior. We
don't bother validating the older behavior as this is standard *nix
behavior.

Signed-off-by: Stephen Finucane <stephen@that.guru>