docs: Split into multiple pages
This mostly reverts commit 4c8deae1. We now have enough documentation to
warrant separate pages and the RTD theme works better with them. We keep
the "Usage" and "Release Notes" sections so any anchors pointing here
are not broken (though the same cannot be said of any links to
subcommands).

The copyright year is updated.

Signed-off-by: Stephen Finucane <stephen@that.guru>