Re-add support for filtering by submitter name, user email
With v1.0 of the Patchwork API, we did not have the ability to directly
filter (server-side) by email address or name so we made an initial
query to the '/people' resource to find the matching submitter and
extract the ID. With v1.1, we added support for searching by email
address but not name. We don't want to add server-side name filtering
because it's not unique. However, we can continue to do this client-side
and fix the regression in behavior.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Closes: #24