Don't exit if multiple matches found for a filter
Previously, specifying a filter that returned multiple results would
result in an error. This made sense for API v1.0 because filters could
only be specified once. However, this is not necessarily the case for
v1.1, which does support multiple filters. In addition, it was
inconsistent for v1.0, as specifying multiple filters on the command
line would result in a warning, not an error.

Resolve this by only warning the user if multiple matches are found for
a given filter. This requires adding a check to ensure at least three
characters are provided for the filter so as not to make it too generic.

Signed-off-by: Stephen Finucane <stephen@that.guru>