series: Apply patches one by one
Instead of downloading an entire series and applying an entire
patch-set, download individual patches and then apply them
one by one using git-am.

This is really useful when patches in a series do not apply cleanly,
and git rejects one of them. In this case, git might apply some
of the patches, and then allow the user to take care of the one that
was rejected.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>