Add spec file
I want to be able to install 'git-pw' from the Fedora repos. However,
all my experiments to date have attempted to use pyp2rpm and have failed.
Add our own handwritten specfile, which is mostly the output of 'pyp2rpm
git-pw' but with the description modified to strip out rST noise, the
man pages installed and the requirements modified to use ranges instead
of the compatible operator (which RPM doesn't support).

Signed-off-by: Stephen Finucane <stephen@that.guru>