CI: Stop publishing to test.pypi.org
The idea behind this is to build packages for every commit and ensure
things never break. However, we're racing because I typically push the
branch followed shortly thereafter by the tag. By time we get to
building the package, the tag is available which means pbr is using the
next release. However, the job wasn't started by the tag push so we
don't build the actual release, leaving that until a later job.

We don't really need this as we can always cut a new release later, so
simply skip it.

Signed-off-by: Stephen Finucane <stephen@that.guru>