Add missing runtime dependencies
We removed these in commit 5a3341d2f34 based on suggestions from a
review [1], however, it seems this was either bad guidance or the magic
sauce needed for this is not present in Copr (or locally, for that
matter). Since we removed these, we've dropped a dependency (six) and
added a new one (pyyaml) so things are slightly different.

Signed-off-by: Stephen Finucane <stephen@that.guru>